requirements.txt
mysql-connector-python==8.0.15
openpyxl==2.6.2